Refactoring `eth2_client` tests
