+++ contracts/near/eth2-client/src/tests/mod.rs
@@ -1,0 +1,3 @@
+pub mod integration_tests;
+pub mod unit_tests;
+pub mod utils;
